moving data to an outer layer
