+++ app/services/sales-api/main.go
@@ -15,7 +15,7 @@ import (
 	"github.com/ardanlabs/conf/v2"
 	"github.com/ardanlabs/service/app/services/sales-api/handlers"
 	"github.com/ardanlabs/service/business/sys/auth"
-	"github.com/ardanlabs/service/business/sys/database"
+	"github.com/ardanlabs/service/data/database"
 	"github.com/ardanlabs/service/foundation/keystore"
 	"github.com/ardanlabs/service/foundation/logger"
 	"go.opentelemetry.io/otel"
