+++ app/tooling/sales-admin/commands/seed.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/ardanlabs/service/business/data/dbschema"
-	"github.com/ardanlabs/service/business/sys/database"
+	"github.com/ardanlabs/service/data/database"
+	"github.com/ardanlabs/service/data/dbschema"
 )
 
 // Seed loads test data into the database.
