+++ app/tooling/sales-admin/main.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/ardanlabs/conf/v2"
 	"github.com/ardanlabs/service/app/tooling/sales-admin/commands"
-	"github.com/ardanlabs/service/business/sys/database"
+	"github.com/ardanlabs/service/data/database"
 	"github.com/ardanlabs/service/foundation/logger"
 	"go.uber.org/zap"
 )
